requirements.txt
paramiko~=3.2.0
python-decouple~=3.8
psycopg2~=2.9.5
SQLAlchemy~=2.0.18